update color component
